adding compiler flags and fixing compiler warnings

+++ src/functions.cpp
@@ -5,7 +5,7 @@
 #include <algorithm>
 
 void print(std::vector<int> const & numbers) {
-    for (int i = 0; i < numbers.size(); ++i) {
+    for (size_t i = 0; i < numbers.size(); ++i) {
         std::cout << i << ": " << numbers[i] << '\n';
     }
     std::cout << '\n';
